PIM-4238: Added back factory family
