+++ src/Pim/Bundle/CatalogBundle/Resources/config/builders.yml
@@ -23,10 +23,3 @@ services:
             - '@pim_catalog.builder.product'
             - %pim_catalog.entity.product_template.class%
             - %pim_catalog.entity.product.class%
-
-    pim_catalog.builder.family:
-        class: %pim_catalog.builder.family.class%
-        arguments:
-            - '@pim_catalog.repository.channel'
-            - '@pim_catalog.factory.attribute_requirement'
-            - '@pim_catalog.repository.attribute'
