+++ src/Pim/Bundle/CatalogBundle/Resources/config/updaters.yml
@@ -116,7 +116,10 @@ services:
         class: %pim_catalog.updater.family.class%
         arguments:
             - '@pim_catalog.repository.family'
-            - '@pim_catalog.builder.family'
+            - '@pim_catalog.factory.family'
+            - '@pim_catalog.repository.attribute'
+            - '@pim_catalog.repository.channel'
+            - '@pim_catalog.factory.attribute_requirement'
 
     pim_catalog.updater.association_type:
         class: %pim_catalog.updater.association_type.class%
