+++ src/Pim/Bundle/TransformBundle/Transformer/FamilyTransformer.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\TransformBundle\Transformer;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
-use Pim\Bundle\CatalogBundle\Builder\FamilyBuilderInterface;
+use Pim\Bundle\CatalogBundle\Factory\FamilyFactory;
 use Pim\Bundle\CatalogBundle\Model\FamilyInterface;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoTransformerInterface;
 use Pim\Bundle\TransformBundle\Transformer\Guesser\GuesserInterface;
@@ -19,10 +19,14 @@ use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
  */
 class FamilyTransformer extends NestedEntityTransformer
 {
-    /** @var FamilyBuilderInterface */
-    protected $familyBuilder;
+    /**
+     * @var FamilyFactory
+     */
+    protected $factory;
 
-    /** @var string */
+    /**
+     * @var string
+     */
     protected $requirementClass;
 
     /**
@@ -33,7 +37,7 @@ class FamilyTransformer extends NestedEntityTransformer
      * @param GuesserInterface               $guesser
      * @param ColumnInfoTransformerInterface $colInfoTransformer
      * @param EntityTransformerInterface     $transformerRegistry
-     * @param FamilyBuilderInterface         $familyBuilder
+     * @param FamilyFactory                  $factory
      * @param string                         $requirementClass
      */
     public function __construct(
@@ -42,11 +46,11 @@ class FamilyTransformer extends NestedEntityTransformer
         GuesserInterface $guesser,
         ColumnInfoTransformerInterface $colInfoTransformer,
         EntityTransformerInterface $transformerRegistry,
-        FamilyBuilderInterface $familyBuilder,
+        FamilyFactory $factory,
         $requirementClass
     ) {
         parent::__construct($doctrine, $propertyAccessor, $guesser, $colInfoTransformer, $transformerRegistry);
-        $this->familyBuilder = $familyBuilder;
+        $this->factory = $factory;
         $this->requirementClass = $requirementClass;
     }
 
@@ -55,7 +59,7 @@ class FamilyTransformer extends NestedEntityTransformer
      */
     protected function createEntity($class, array $data)
     {
-        return $this->familyBuilder->createFamily(true);
+        return $this->factory->createFamily();
     }
 
     /**
